Move tests to alternate file
