+++ tests/test_validate.py
@@ -23,8 +23,6 @@ from pointblank.validate import (
     _fmt_lg,
     _create_table_time_html,
     _create_table_type_html,
-    _select_df_lib,
-    _get_tbl_type,
 )
 from pointblank.thresholds import Thresholds
 from pointblank.schema import Schema
@@ -1938,37 +1936,6 @@ def test_create_table_type_html():
     assert _create_table_type_html(
         tbl_type="pandas", tbl_name="tbl_name"
     ) != _create_table_type_html(tbl_type="pandas", tbl_name=None)
-
-
-def test_select_df_lib():
-
-    # Mock the absence of the both the Pandas and Polars libraries
-    with patch.dict(sys.modules, {"pandas": None, "polars": None}):
-        # An ImportError is raised when the `pandas` and `polars` packages are not installed
-        with pytest.raises(ImportError):
-            _select_df_lib()
-
-    # Mock the absence of the Pandas library
-    with patch.dict(sys.modules, {"pandas": None}):
-        # The Polars library is selected when the `pandas` package is not installed
-        assert _select_df_lib(preference="polars") == pl
-        assert _select_df_lib(preference="pandas") == pl
-
-    # Mock the absence of the Polars library
-    with patch.dict(sys.modules, {"polars": None}):
-        # The Pandas library is selected when the `polars` package is not installed
-        assert _select_df_lib(preference="pandas") == pd
-        assert _select_df_lib(preference="polars") == pd
-
-    # Where both the Pandas and Polars libraries are available
-    assert _select_df_lib(preference="pandas") == pd
-    assert _select_df_lib(preference="polars") == pl
-
-
-def test_get_tbl_type():
-
-    assert _get_tbl_type(pd.DataFrame()) == "pandas"
-    assert _get_tbl_type(pl.DataFrame()) == "polars"
 
 
 def test_pointblank_config_class():
